requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.1
PyJWT==2.10.1
Pillow==10.4.0
django-environ==0.12.0
django-cors-headers==4.9.0
mysqlclient==2.2.7
django-modeltranslation==0.19.17
drf-spectacular==0.28.0
